+++ tests/test_version.py
@@ -2,4 +2,4 @@ from imr_proxy.version import get_version
 
 
 def test_version():
-    assert get_version() == "0.1.82"
+    assert get_version() == "0.1.83"